// app/types/sportType.ts
/**
 * Type `SportType` définissant les propriétés d'une épreuve sportive.
 *
 * @property id_sport - Identifiant unique de l'épreuve sportive.
 * @property title - Nom de l'épreuve sportive.
 * @property image - Url de l'image représentant l'épreuve sportive.
 */
export type SportType = {
  
  id_sport: number
  title: string
  image: string
}